requirements.txt
openai
langchain==0.1.0
langchain_openai==0.0.2
langchainhub==0.1.14

# Agent tools
google-search-results
requests

# Qualirt of life tooling
python-dotenv